Split BsqSwap tests into "buy" and "sell" tests
