+++ apitest/src/test/java/bisq/apitest/method/trade/BsqSwapBuyBtcTradeTest.java
@@ -17,17 +17,16 @@
 
 package bisq.apitest.method.trade;
 
-import bisq.proto.grpc.OfferInfo;
 import bisq.proto.grpc.TradeInfo;
 
-import java.util.ArrayList;
-import java.util.List;
+import protobuf.OfferDirection;
 
 import lombok.Setter;
 import lombok.extern.slf4j.Slf4j;
 
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.MethodOrderer;
 import org.junit.jupiter.api.Order;
 import org.junit.jupiter.api.Test;
@@ -35,6 +34,7 @@ import org.junit.jupiter.api.TestMethodOrder;
 
 import static bisq.apitest.config.ApiTestConfig.BSQ;
 import static bisq.apitest.config.ApiTestConfig.BTC;
+import static bisq.core.offer.OfferDirection.BUY;
 import static bisq.proto.grpc.GetOfferCategoryReply.OfferCategory.BSQ_SWAP;
 import static java.lang.String.format;
 import static org.junit.jupiter.api.Assertions.assertEquals;
@@ -43,17 +43,16 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 import static org.junit.jupiter.api.Assertions.fail;
 import static protobuf.BsqSwapTrade.State.COMPLETED;
 import static protobuf.BsqSwapTrade.State.PREPARATION;
-import static protobuf.OfferDirection.BUY;
 
 
 
 import bisq.apitest.method.offer.AbstractOfferTest;
 import bisq.cli.GrpcClient;
 
-// @Disabled
+@Disabled
 @Slf4j
 @TestMethodOrder(MethodOrderer.OrderAnnotation.class)
-public class BsqSwapTradeTest extends AbstractTradeTest {
+public class BsqSwapBuyBtcTradeTest extends AbstractTradeTest {
 
     // Long-running swap trade tests might want to check node logs for exceptions.
     @Setter
@@ -80,15 +79,16 @@ public class BsqSwapTradeTest extends AbstractTradeTest {
 
     @Test
     @Order(2)
-    public void testAliceCreateBsqSwapBuyOffer() {
-        var mySwapOffer = aliceClient.createBsqSwapOffer(BUY.name(),
+    public void testAliceCreateBsqSwapBuyBtcOffer() {
+        // Alice buys BTC, pays trade fee.  Bob (BTC seller) pays miner tx fee.
+        var mySwapOffer = aliceClient.createBsqSwapOffer(OfferDirection.BUY.name(),
                 1_000_000L,     // 0.01 BTC
                 1_000_000L,
                 "0.00005");
         log.debug("Pending BsqSwap Sell BSQ (Buy BTC) OFFER:\n{}", toOfferTable.apply(mySwapOffer));
         var newOfferId = mySwapOffer.getId();
         assertNotEquals("", newOfferId);
-        assertEquals(BUY.name(), mySwapOffer.getDirection());
+        assertEquals(OfferDirection.BUY.name(), mySwapOffer.getDirection());
         assertEquals(5_000, mySwapOffer.getPrice());
         assertEquals(1_000_000L, mySwapOffer.getAmount());
         assertEquals(1_000_000L, mySwapOffer.getMinAmount());
@@ -107,7 +107,7 @@ public class BsqSwapTradeTest extends AbstractTradeTest {
     @Test
     @Order(3)
     public void testBobTakesBsqSwapOffer() {
-        var availableSwapOffer = getAvailableBsqSwapOffer(bobClient);
+        var availableSwapOffer = getAvailableBsqSwapOffer(bobClient, BUY, true);
 
         // Before sending a TakeOfferRequest, the CLI needs to know what kind of Offer
         // it is taking (v1 or BsqSwap).  Only BSQ swap offers can be taken with a
@@ -161,32 +161,6 @@ public class BsqSwapTradeTest extends AbstractTradeTest {
         log.debug("Alice's After Trade Balance:\n{}", formatBalancesTbls(alicesBalances));
         var bobsBalances = bobClient.getBalances();
         log.debug("Bob's After Trade Balance:\n{}", formatBalancesTbls(bobsBalances));
-    }
-
-    private OfferInfo getAvailableBsqSwapOffer(GrpcClient client) {
-        List<OfferInfo> bsqSwapOffers = new ArrayList<>();
-        int numFetchAttempts = 0;
-        while (bsqSwapOffers.size() == 0) {
-            bsqSwapOffers.addAll(client.getBsqSwapOffers(BUY.name()));
-            numFetchAttempts++;
-            if (bsqSwapOffers.size() == 0) {
-                log.warn("No available bsq swap offers found after {} fetch attempts.", numFetchAttempts);
-                if (numFetchAttempts > 9) {
-                    if (checkForLoggedExceptions) {
-                        printNodeExceptionMessages(log);
-                    }
-                    fail(format("Bob gave up on fetching available bsq swap offers after %d attempts.", numFetchAttempts));
-                }
-                sleep(1_000);
-            } else {
-                assertEquals(1, bsqSwapOffers.size());
-                log.debug("Bob found new available bsq swap offer on attempt # {}.", numFetchAttempts);
-                break;
-            }
-        }
-        var bsqSwapOffer = bobClient.getBsqSwapOffer(bsqSwapOffers.get(0).getId());
-        assertEquals(bsqSwapOffers.get(0).getId(), bsqSwapOffer.getId());
-        return bsqSwapOffer;
     }
 
     private TradeInfo getBsqSwapTrade(GrpcClient client, String tradeId) {
